Add files via upload
The results of a survey of the households in your township have been made available. Each record contains data for one household, including a four digit integer identification number, the annual income of the household and the number of members in the family. You may assume that no more than 10 households were surveyed. Devise a C/C++ program using structures and Functions to implement the following.

(a) Calculate the average household income. List the identification number and income of each household whose income exceeds the average.  
(b) Sort the survey list based on the number of members in the family. (Smallest to largest)